Code refactoring, removed secondary test class
