+++ tests/ImageSharp.Tests/Processing/Processors/Effects/PixelShaderTest.cs
@@ -6,13 +6,13 @@ using System.Numerics;
 
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing;
-
+using SixLabors.ImageSharp.Processing.Processors.Effects;
 using Xunit;
 
 namespace SixLabors.ImageSharp.Tests.Processing.Processors.Effects
 {
     [GroupOutput("Effects")]
-    public partial class PixelShaderTest
+    public class PixelShaderTest
     {
         [Theory]
         [WithFile(TestImages.Png.CalliphoraPartial, PixelTypes.Rgba32)]
@@ -21,15 +21,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Processors.Effects
         {
             provider.RunValidatingProcessorTest(
                 x => x.ProcessPixelRowsAsVector4(
-                span =>
-                {
-                    for (int i = 0; i < span.Length; i++)
-                    {
-                        Vector4 v4 = span[i];
-                        float avg = (v4.X + v4.Y + v4.Z) / 3f;
-                        span[i] = new Vector4(avg);
-                    }
-                }),
+                span => default(PixelAverageProcessor).Invoke(span)),
                 appendPixelTypeToFileName: false);
         }
 
@@ -40,15 +32,20 @@ namespace SixLabors.ImageSharp.Tests.Processing.Processors.Effects
         {
             provider.RunRectangleConstrainedValidatingProcessorTest(
                 (x, rect) => x.ProcessPixelRowsAsVector4(
-                    span =>
-                    {
-                        for (int i = 0; i < span.Length; i++)
-                        {
-                            Vector4 v4 = span[i];
-                            float avg = (v4.X + v4.Y + v4.Z) / 3f;
-                            span[i] = new Vector4(avg);
-                        }
-                    }, rect));
+                    span => default(PixelAverageProcessor).Invoke(span), rect));
+        }
+
+        private readonly struct PixelAverageProcessor : IPixelRowDelegate
+        {
+            public void Invoke(Span<Vector4> span)
+            {
+                for (int i = 0; i < span.Length; i++)
+                {
+                    Vector4 v4 = span[i];
+                    float avg = (v4.X + v4.Y + v4.Z) / 3f;
+                    span[i] = new Vector4(avg);
+                }
+            }
         }
 
         [Theory]
@@ -58,26 +55,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Processors.Effects
         {
             provider.RunValidatingProcessorTest(
                 c => c.ProcessPositionAwarePixelRowsAsVector4(
-                    (span, offset) =>
-                    {
-                        int y = offset.Y;
-                        int x = offset.X;
-                        for (int i = 0; i < span.Length; i++)
-                        {
-                            float
-                                sine = MathF.Sin(y),
-                                cosine = MathF.Cos(x + i),
-                                sum = sine + cosine,
-                                abs = MathF.Abs(sum),
-                                a = 0.5f + (abs / 2); // Max value for sin(y) + cos(x) is 2
-
-                            Vector4 v4 = span[i];
-                            float avg = (v4.X + v4.Y + v4.Z) / 3f;
-                            var gray = new Vector4(avg, avg, avg, a);
-
-                            span[i] = Vector4.Clamp(gray, Vector4.Zero, Vector4.One);
-                        }
-                    }),
+                    (span, offset) => default(TrigonometryProcessor).Invoke(span, offset)),
                 appendPixelTypeToFileName: false);
         }
 
@@ -88,26 +66,31 @@ namespace SixLabors.ImageSharp.Tests.Processing.Processors.Effects
         {
             provider.RunRectangleConstrainedValidatingProcessorTest(
                 (c, rect) => c.ProcessPositionAwarePixelRowsAsVector4(
-                    (span, offset) =>
-                    {
-                        int y = offset.Y;
-                        int x = offset.X;
-                        for (int i = 0; i < span.Length; i++)
-                        {
-                            float
-                                sine = MathF.Sin(y),
-                                cosine = MathF.Cos(x + i),
-                                sum = sine + cosine,
-                                abs = MathF.Abs(sum),
-                                a = 0.5f + (abs / 2);
+                    (span, offset) => default(TrigonometryProcessor).Invoke(span, offset), rect));
+        }
 
-                            Vector4 v4 = span[i];
-                            float avg = (v4.X + v4.Y + v4.Z) / 3f;
-                            var gray = new Vector4(avg, avg, avg, a);
+        private readonly struct TrigonometryProcessor : IPixelRowDelegate<Point>
+        {
+            public void Invoke(Span<Vector4> span, Point value)
+            {
+                int y = value.Y;
+                int x = value.X;
+                for (int i = 0; i < span.Length; i++)
+                {
+                    float
+                        sine = MathF.Sin(y),
+                        cosine = MathF.Cos(x + i),
+                        sum = sine + cosine,
+                        abs = MathF.Abs(sum),
+                        a = 0.5f + (abs / 2);
 
-                            span[i] = Vector4.Clamp(gray, Vector4.Zero, Vector4.One);
-                        }
-                    }, rect));
+                    Vector4 v4 = span[i];
+                    float avg = (v4.X + v4.Y + v4.Z) / 3f;
+                    var gray = new Vector4(avg, avg, avg, a);
+
+                    span[i] = Vector4.Clamp(gray, Vector4.Zero, Vector4.One);
+                }
+            }
         }
     }
 }
